准备测试界面展示数据

+++ src/MCCS.Collecter/HardwareDevices/DataPoint.cs
@@ -9,7 +9,7 @@
 
     public record DataPoint
     {
-        public int DeviceId { get; init; } 
+        public long DeviceId { get; init; } 
         public long Timestamp { get; init; }
         public object Value { get; init; }
         public string Unit { get; init; }
